requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.34.0
azure-storage-blob==12.25.1
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
dnspython==2.7.0
idna==3.10
isodate==0.7.2
msrest==0.7.1
oauthlib==3.2.2
pillow==11.2.1
pycparser==2.22
pymongo==4.12.1
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
